fixed up TODO items from re-org.
