+++ README.md
@@ -7,9 +7,9 @@ Pandas tables with the performance results of all the models can be built by:
 ```python
     performance_df, performance_curves_dict = \
         btc.just_benchmark(X_train, y_train, X_test, y_test, 2, classifiers,
-                           STD_BINARY_LOSS, STD_BINARY_CURVES, ref_method)
+                           STD_CLASS_LOSS, STD_BINARY_CURVES, ref_method)
 ```
-This benchmarks all the models in classifiers on the data (`X_train`, `y_train`, `X_test`, `y_test`) for 2-class classification. It uses the loss function described in the dictionaries `STD_BINARY_LOSS`, and the curves (e.g., ROC, PR) in `STD_BINARY_CURVES`. `ref_method` defines the model that is the reference to compare against for assessing statistically significant performance gains.
+This benchmarks all the models in classifiers on the data (`X_train`, `y_train`, `X_test`, `y_test`) for 2-class classification. It uses the loss function described in the dictionaries `STD_CLASS_LOSS`, and the curves (e.g., ROC, PR) in `STD_BINARY_CURVES`. `ref_method` defines the model that is the reference to compare against for assessing statistically significant performance gains.
 
 The sciprint module formats these tables for scientific presentation. The performance dictionaries can be converted to cleanly formatted tables: correct significant figures, shifting of exponent for compactness, thresholding huge/small (crap limit) results, and correct alignment of decimal points, units in headers, etc. Here we use:
 ```python
