+++ data_splitter.py
@@ -321,7 +321,7 @@ def split_df(df, splits=DEFAULT_SPLIT, assume_unique=(), assume_sorted=()):
                                 assume_sorted=(feature in assume_sorted),
                                 assume_unique=(feature in assume_unique))
 
-        # TODO assert dtype bool before ~ operation
+        assert(train_curr.dtype.kind == 'b')  # Make sure ~ does right thing
         train_series &= train_curr
         test_series &= ~train_curr
     assert(not (train_series & test_series).any())
