+++ perf_curves.py
@@ -3,40 +3,9 @@ import numpy as np
 
 EPSILON = 1e-10  # Size of pseudo-point to add to true/false positive count.
 
-
-def make_into_step(xp, yp):
-    """Make pairs of `xp` and `yp` vectors into proper step function. That is,
-    remove NaN `xp` values and multiple steps at same location.
-
-    Parameters
-    ----------
-    xp : 1d np array
-        The sample points corresponding to the y values. Must be sorted.
-    yp : 1d np array
-        Values in y-axis for step function.
-
-    Returns
-    -------
-    xp : 1d np array
-        Input `xp` after removing extra points.
-    yp : 1d np array
-        Input `yp` after removing extra points.
-
-    Notes
-    -----
-    Keeps last value in list when multiple steps happen at the same x-value.
-    """
-    # TODO check if this used, remove if not
-    assert(xp.ndim == 1 and xp.shape == yp.shape)
-
-    idx = ~np.isnan(xp)
-    xp, yp = xp[idx], yp[idx]
-
-    deltas = np.diff(xp)
-    assert(not np.any(deltas < -1e-10))
-
-    idx = [] if xp.size == 0 else np.concatenate((deltas > 0, [True]))
-    return xp[idx], yp[idx]
+# ============================================================================
+# Create general binary count curves
+# ============================================================================
 
 
 def add_pseudo_points(fps, tps):
@@ -157,6 +126,10 @@ def _binary_clf_curve(y_true, y_score, sample_weight=None):
 
     return fps, tps, thresholds
 
+# ============================================================================
+# Convert general binary count curves to ROC, PR, PRG
+# ============================================================================
+
 
 def roc_curve(y_true, y_score, sample_weight=None):
     """Compute ROC curve with optional sample weight matrix.
@@ -196,7 +169,9 @@ def roc_curve(y_true, y_score, sample_weight=None):
 
 
 def recall_precision_curve(y_true, y_score, sample_weight=None):
-    """Compute recall precision curve with optional sample weight matrix.
+    """Compute recall precision curve with optional sample weight matrix. This
+    has intentionally been named recall-precision rather than the traditional
+    precision-recall.
 
     Based on `sklearn.metrics.ranking.precision_recall_curve` except that it
     supports a matrix a different sample weights `sample_weight`. The name
@@ -281,6 +256,10 @@ def prg_curve(y_true, y_score, sample_weight=None):
     assert(np.all((rec_gain == 0.0) | (prec_gain <= 1.0)))
     return rec_gain, prec_gain, thresholds
 
+# ============================================================================
+# Routines for area calculation
+# ============================================================================
+
 
 def auc_trapz(x_curve, y_curve):
     """Compute area under function using trapezoid rule in vectorized way.
@@ -324,113 +303,3 @@ def auc_left(x_curve, y_curve):
     with np.errstate(invalid='ignore'):
         auc = np.nansum(y_curve[:-1, :] * np.diff(x_curve, axis=0), axis=0)
     return auc
-
-
-def _nv_add_pseudo_points(fps, tps):
-    """Non-vectorized version that should be moved off into test file."""
-    if fps[-1] == 0:
-        fps = EPSILON * tps
-        tps = tps.astype(fps.dtype)
-
-    if tps[-1] == 0:
-        tps = EPSILON * fps
-        fps = fps.astype(tps.dtype)
-    return fps, tps
-
-
-def _nv_binary_clf_curve(y_true, y_score, sample_weight=None):
-    """Non-vectorized version that should be moved off into test file."""
-    assert(y_true.ndim == 1 and y_true.dtype.kind == 'b')
-    assert(y_score.shape == y_true.shape and np.all(np.isfinite(y_score)))
-    assert(y_true.size >= 1)
-
-    # sort scores and corresponding truth values
-    desc_score_indices = np.argsort(y_score, kind='mergesort')[::-1]
-    y_score, y_true = y_score[desc_score_indices], y_true[desc_score_indices]
-
-    # y_score typically has many tied values. Here we extract
-    # the indices associated with the distinct values. We also
-    # concatenate a value for the end of the curve.
-    distinct_value_indices = np.where(np.diff(y_score))[0]
-    threshold_idxs = np.r_[distinct_value_indices, y_true.size - 1]
-
-    if sample_weight is None:
-        tps = np.cumsum(y_true)[threshold_idxs]
-        fps = 1 + threshold_idxs - tps
-        assert(fps[-1] == np.sum(~y_true) and tps[-1] == np.sum(y_true))
-    else:
-        assert(sample_weight.shape == y_true.shape)
-        assert(np.all(np.isfinite(sample_weight)))
-        # Negative weight makes no sense, 0 can violate assumps. of other funcs
-        assert(np.all(sample_weight > 0))
-
-        weight = sample_weight[desc_score_indices]
-        tps = np.cumsum(y_true * weight)[threshold_idxs]
-        fps = np.cumsum(weight)[threshold_idxs] - tps
-        assert(np.allclose((fps[-1], tps[-1]),
-                           (np.sum(weight[~y_true]), np.sum(weight[y_true]))))
-
-    # Now put in the (0, 0) coord (y_score >= np.inf)
-    assert(not (tps[0] == 0 and fps[0] == 0))
-    fps, tps = np.r_[0, fps], np.r_[0, tps]
-    thresholds = np.r_[np.inf, y_score[threshold_idxs]]
-
-    # Clean up corner case
-    fps, tps = _nv_add_pseudo_points(fps, tps)
-    assert(fps[-1] > 0 and tps[-1] > 0)
-    assert(fps.dtype == tps.dtype)
-
-    # Remove any decreases due to numerics
-    fps = np.maximum.accumulate(fps)
-    assert(np.all((np.diff(fps) >= 0.0) & (np.diff(tps) >= 0.0)))
-    return fps, tps, thresholds
-
-
-def _nv_roc_curve(y_true, y_score, sample_weight=None):
-    """Non-vectorized version that should be moved off into test file."""
-    fps, tps, thresholds = _nv_binary_clf_curve(y_true, y_score,
-                                                sample_weight=sample_weight)
-    fpr = np.true_divide(fps, fps[-1])
-    tpr = np.true_divide(tps, tps[-1])
-    return fpr, tpr, thresholds
-
-
-def _nv_recall_precision_curve(y_true, y_score, sample_weight=None):
-    """Non-vectorized version that should be moved off into test file."""
-    fps, tps, thresholds = _nv_binary_clf_curve(y_true, y_score,
-                                                sample_weight=sample_weight)
-    recall = np.true_divide(tps, tps[-1])
-    with np.errstate(divide='ignore', invalid='ignore'):
-        precision = np.true_divide(tps, tps + fps)
-    precision[0] = precision[1]
-    assert(np.all(0.0 <= precision) and np.all(precision <= 1.0))
-    return recall, precision, thresholds
-
-
-def _nv_prg_curve(y_true, y_score, sample_weight=None):
-    """Non-vectorized version that should be moved off into test file."""
-    fps, tps, thresholds = _nv_binary_clf_curve(y_true, y_score,
-                                                sample_weight=sample_weight)
-    n_neg, n_pos = fps[-1], tps[-1]
-    fns = n_pos - tps
-
-    den = n_neg * tps
-    with np.errstate(divide='ignore', invalid='ignore'):
-        rec_gain = 1.0 - np.true_divide(n_pos * fns, den)
-        prec_gain = 1.0 - np.true_divide(n_pos * fps, den)
-    # interpolate backward just like in PR curve
-    prec_gain[0] = prec_gain[1]
-    assert(np.all(rec_gain <= 1.0) and np.all(prec_gain <= 1.0))
-
-    # Find index to put everything in the box
-    with np.errstate(invalid='ignore'):
-        assert(not np.any(np.diff(rec_gain) < 0.0))
-    idx = np.searchsorted(rec_gain, 0.0, side='right')
-    assert(idx == np.where(rec_gain > 0.0)[0][0])
-    assert(idx > 0)  # Not selecting first point
-
-    # Bring forward most recent negative point as point at 0
-    rec_gain = np.concatenate(([0.0], rec_gain[idx:]))
-    prec_gain = prec_gain[idx - 1:]
-    thresholds = thresholds[idx - 1:]
-    return rec_gain, prec_gain, thresholds
