+++ perf_curves_test.py
@@ -5,20 +5,118 @@ from sklearn.metrics.ranking import _binary_clf_curve
 from sklearn.metrics.ranking import roc_curve, precision_recall_curve
 import perf_curves as pc
 
+# ============================================================================
+# Non-vectorized versions of routines in perf_curves for testing.
+# ============================================================================
 
-def eval_step_func_test():
-    N = np.random.randint(low=0, high=10)
 
-    xp = np.sort(np.random.choice(np.random.randn(N + 1),
-                                  size=N, replace=True))
-    yp = np.random.randn(N)
-    D = {xp[ii]: yp[ii] for ii in xrange(N)}
+def _nv_add_pseudo_points(fps, tps):
+    if fps[-1] == 0:
+        fps = pc.EPSILON * tps
+        tps = tps.astype(fps.dtype)
 
-    xp2, yp2 = pc.make_into_step(xp, yp)
-    assert(xp2.shape == yp2.shape)
-    assert(np.all(np.unique(xp) == xp2))
-    D2 = {xp2[ii]: yp2[ii] for ii in xrange(len(xp2))}
-    assert(D == D2)
+    if tps[-1] == 0:
+        tps = pc.EPSILON * fps
+        fps = fps.astype(tps.dtype)
+    return fps, tps
+
+
+def _nv_binary_clf_curve(y_true, y_score, sample_weight=None):
+    assert(y_true.ndim == 1 and y_true.dtype.kind == 'b')
+    assert(y_score.shape == y_true.shape and np.all(np.isfinite(y_score)))
+    assert(y_true.size >= 1)
+
+    # sort scores and corresponding truth values
+    desc_score_indices = np.argsort(y_score, kind='mergesort')[::-1]
+    y_score, y_true = y_score[desc_score_indices], y_true[desc_score_indices]
+
+    # y_score typically has many tied values. Here we extract
+    # the indices associated with the distinct values. We also
+    # concatenate a value for the end of the curve.
+    distinct_value_indices = np.where(np.diff(y_score))[0]
+    threshold_idxs = np.r_[distinct_value_indices, y_true.size - 1]
+
+    if sample_weight is None:
+        tps = np.cumsum(y_true)[threshold_idxs]
+        fps = 1 + threshold_idxs - tps
+        assert(fps[-1] == np.sum(~y_true) and tps[-1] == np.sum(y_true))
+    else:
+        assert(sample_weight.shape == y_true.shape)
+        assert(np.all(np.isfinite(sample_weight)))
+        # Negative weight makes no sense, 0 can violate assumps. of other funcs
+        assert(np.all(sample_weight > 0))
+
+        weight = sample_weight[desc_score_indices]
+        tps = np.cumsum(y_true * weight)[threshold_idxs]
+        fps = np.cumsum(weight)[threshold_idxs] - tps
+        assert(np.allclose((fps[-1], tps[-1]),
+                           (np.sum(weight[~y_true]), np.sum(weight[y_true]))))
+
+    # Now put in the (0, 0) coord (y_score >= np.inf)
+    assert(not (tps[0] == 0 and fps[0] == 0))
+    fps, tps = np.r_[0, fps], np.r_[0, tps]
+    thresholds = np.r_[np.inf, y_score[threshold_idxs]]
+
+    # Clean up corner case
+    fps, tps = _nv_add_pseudo_points(fps, tps)
+    assert(fps[-1] > 0 and tps[-1] > 0)
+    assert(fps.dtype == tps.dtype)
+
+    # Remove any decreases due to numerics
+    fps = np.maximum.accumulate(fps)
+    assert(np.all((np.diff(fps) >= 0.0) & (np.diff(tps) >= 0.0)))
+    return fps, tps, thresholds
+
+
+def _nv_roc_curve(y_true, y_score, sample_weight=None):
+    fps, tps, thresholds = _nv_binary_clf_curve(y_true, y_score,
+                                                sample_weight=sample_weight)
+    fpr = np.true_divide(fps, fps[-1])
+    tpr = np.true_divide(tps, tps[-1])
+    return fpr, tpr, thresholds
+
+
+def _nv_recall_precision_curve(y_true, y_score, sample_weight=None):
+    fps, tps, thresholds = _nv_binary_clf_curve(y_true, y_score,
+                                                sample_weight=sample_weight)
+    recall = np.true_divide(tps, tps[-1])
+    with np.errstate(divide='ignore', invalid='ignore'):
+        precision = np.true_divide(tps, tps + fps)
+    precision[0] = precision[1]
+    assert(np.all(0.0 <= precision) and np.all(precision <= 1.0))
+    return recall, precision, thresholds
+
+
+def _nv_prg_curve(y_true, y_score, sample_weight=None):
+    fps, tps, thresholds = _nv_binary_clf_curve(y_true, y_score,
+                                                sample_weight=sample_weight)
+    n_neg, n_pos = fps[-1], tps[-1]
+    fns = n_pos - tps
+
+    den = n_neg * tps
+    with np.errstate(divide='ignore', invalid='ignore'):
+        rec_gain = 1.0 - np.true_divide(n_pos * fns, den)
+        prec_gain = 1.0 - np.true_divide(n_pos * fps, den)
+    # interpolate backward just like in PR curve
+    prec_gain[0] = prec_gain[1]
+    assert(np.all(rec_gain <= 1.0) and np.all(prec_gain <= 1.0))
+
+    # Find index to put everything in the box
+    with np.errstate(invalid='ignore'):
+        assert(not np.any(np.diff(rec_gain) < 0.0))
+    idx = np.searchsorted(rec_gain, 0.0, side='right')
+    assert(idx == np.where(rec_gain > 0.0)[0][0])
+    assert(idx > 0)  # Not selecting first point
+
+    # Bring forward most recent negative point as point at 0
+    rec_gain = np.concatenate(([0.0], rec_gain[idx:]))
+    prec_gain = prec_gain[idx - 1:]
+    thresholds = thresholds[idx - 1:]
+    return rec_gain, prec_gain, thresholds
+
+# ============================================================================
+# Now the actual tests
+# ============================================================================
 
 
 def nv_binary_clf_curve_test():
@@ -36,8 +134,7 @@ def nv_binary_clf_curve_test():
         sample_weight = np.maximum(np.random.multinomial(N, np.ones(N) / N),
                                    1e-6)
 
-    fps, tps, thresholds = pc._nv_binary_clf_curve(y_bool, y_pred,
-                                                   sample_weight)
+    fps, tps, thresholds = _nv_binary_clf_curve(y_bool, y_pred, sample_weight)
     assert(fps.shape == tps.shape and fps.shape == thresholds.shape)
     assert(np.all(np.isfinite(fps)))
     assert(np.all(np.isfinite(tps)))
@@ -53,7 +150,7 @@ def nv_binary_clf_curve_test():
     assert(np.all((np.diff(fps) > 0) | (np.diff(tps) > 0)))
     assert(np.all(np.diff(thresholds) < 0.0))
 
-    fpr, tpr, thresholds_roc = pc._nv_roc_curve(y_bool, y_pred, sample_weight)
+    fpr, tpr, thresholds_roc = _nv_roc_curve(y_bool, y_pred, sample_weight)
     assert(fpr.shape == tpr.shape and fpr.shape == thresholds_roc.shape)
     assert(np.all(np.isfinite(fpr)))
     assert(np.all(np.isfinite(tpr)))
@@ -64,8 +161,8 @@ def nv_binary_clf_curve_test():
     assert(np.all((np.diff(fpr) > 0.0) | (np.diff(tpr) > 0.0)))
     assert(np.all(np.diff(thresholds_roc) < 0.0))
 
-    rec, prec, thresholds_pr = pc._nv_recall_precision_curve(y_bool, y_pred,
-                                                             sample_weight)
+    rec, prec, thresholds_pr = _nv_recall_precision_curve(y_bool, y_pred,
+                                                          sample_weight)
     assert(rec.shape == prec.shape and rec.shape == thresholds_pr.shape)
     assert(np.all(np.isfinite(rec)))
     assert(np.all(np.isfinite(prec)))
@@ -79,8 +176,8 @@ def nv_binary_clf_curve_test():
     assert(np.all(np.diff(rec) >= 0.0))
     assert(np.all(np.diff(thresholds_pr) < 0.0))
 
-    rec_gain, prec_gain, thresholds_prg = pc._nv_prg_curve(y_bool, y_pred,
-                                                           sample_weight)
+    rec_gain, prec_gain, thresholds_prg = _nv_prg_curve(y_bool, y_pred,
+                                                        sample_weight)
     assert(rec_gain.shape == prec_gain.shape)
     assert(rec_gain.shape == thresholds_prg.shape)
     assert(np.all(np.isfinite(thresholds_prg[1:])))
@@ -192,29 +289,28 @@ def binary_clf_curve_test():
 
     if sample_weight is None:
         fps2, tps2, thresholds2 = \
-            pc._nv_binary_clf_curve(y_bool, y_pred, np.ones(N, dtype=int))
+            _nv_binary_clf_curve(y_bool, y_pred, np.ones(N, dtype=int))
         assert(np.all(fps2 == fps[:, 0]))
         assert(np.all(tps2 == tps[:, 0]))
         assert(np.all(thresholds2 == thresholds))
 
-        fps2, tps2, thresholds2 = pc._nv_binary_clf_curve(y_bool, y_pred)
+        fps2, tps2, thresholds2 = _nv_binary_clf_curve(y_bool, y_pred)
         assert(np.all(fps2 == fps[:, 0]))
         assert(np.all(tps2 == tps[:, 0]))
         assert(np.all(thresholds2 == thresholds))
 
-        fpr2, tpr2, thresholds_roc2 = pc._nv_roc_curve(y_bool, y_pred)
+        fpr2, tpr2, thresholds_roc2 = _nv_roc_curve(y_bool, y_pred)
         assert(np.allclose(fpr2, fpr[:, 0]))
         assert(np.allclose(tpr2, tpr[:, 0]))
         assert(np.allclose(thresholds_roc2, thresholds))
 
         rec2, prec2, thresholds_pr2 = \
-            pc._nv_recall_precision_curve(y_bool, y_pred)
+            _nv_recall_precision_curve(y_bool, y_pred)
         assert(np.allclose(rec2, rec[:, 0]))
         assert(np.allclose(prec2, prec[:, 0]))
         assert(np.allclose(thresholds_pr2, thresholds_pr))
 
-        rec_gain2, prec_gain2, thresholds_prg2 = \
-            pc._nv_prg_curve(y_bool, y_pred)
+        rec_gain2, prec_gain2, thresholds_prg2 = _nv_prg_curve(y_bool, y_pred)
         assert(np.all(rec_gain[:-len(rec_gain2), 0] == 0.0))
         assert(np.allclose(rec_gain2, rec_gain[-len(rec_gain2):, 0]))
         assert(np.allclose(prec_gain2, prec_gain[-len(rec_gain2):, 0]))
@@ -225,19 +321,19 @@ def binary_clf_curve_test():
         weight_curr = sample_weight[:, ii]
 
         fpr2, tpr2, thresholds_roc2 = \
-            pc._nv_roc_curve(y_bool, y_pred, weight_curr)
+            _nv_roc_curve(y_bool, y_pred, weight_curr)
         assert(np.allclose(fpr2, fpr[:, ii]))
         assert(np.allclose(tpr2, tpr[:, ii]))
         assert(np.allclose(thresholds_roc2, thresholds))
 
         rec2, prec2, thresholds_pr2 = \
-            pc._nv_recall_precision_curve(y_bool, y_pred, weight_curr)
+            _nv_recall_precision_curve(y_bool, y_pred, weight_curr)
         assert(np.allclose(rec2, rec[:, ii]))
         assert(np.allclose(prec2, prec[:, ii]))
         assert(np.allclose(thresholds_pr2, thresholds_pr))
 
         rec_gain2, prec_gain2, thresholds_prg2 = \
-            pc._nv_prg_curve(y_bool, y_pred, weight_curr)
+            _nv_prg_curve(y_bool, y_pred, weight_curr)
         assert(np.all(rec_gain[:-len(rec_gain2), ii] == 0.0))
         assert(np.allclose(rec_gain2, rec_gain[-len(rec_gain2):, ii]))
         assert(np.allclose(prec_gain2, prec_gain[-len(rec_gain2):, ii]))
@@ -245,9 +341,7 @@ def binary_clf_curve_test():
 
 np.random.seed(89254)
 
-print 'start'
 for rr in xrange(100000):
-    eval_step_func_test()
     nv_binary_clf_curve_test()
     binary_clf_curve_test()
-print 'done'
+print 'passed'
