+++ util_test.py
@@ -65,6 +65,21 @@ def eval_step_func_test():
 
     assert(np.allclose(y_grid, y_grid2))
 
+
+def make_into_step_test():
+    N = np.random.randint(low=0, high=10)
+
+    xp = np.sort(np.random.choice(np.random.randn(N + 1),
+                                  size=N, replace=True))
+    yp = np.random.randn(N)
+    D = {xp[ii]: yp[ii] for ii in xrange(N)}
+
+    xp2, yp2 = util.make_into_step(xp, yp)
+    assert(xp2.shape == yp2.shape)
+    assert(np.all(np.unique(xp) == xp2))
+    D2 = {xp2[ii]: yp2[ii] for ii in xrange(len(xp2))}
+    assert(D == D2)
+
 np.random.seed(75675)
 
 for _ in xrange(1000):
@@ -72,4 +87,5 @@ for _ in xrange(1000):
     test_normalize()
     epsilon_noise_test()
     eval_step_func_test()
+    make_into_step_test()
 print 'passed'
